feat #6: request mapping 이슈 해결 - controller 분리
