+++ .gitignore
@@ -4,7 +4,7 @@ build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
-src/main/resources/security-application.yml
+src/main/resources/application.yml
 
 ### STS ###
 .apt_generated
